Refactored some I/O and Lexer code.
On branch master

	new file:   .gitignore
	modified:   Makefile
	new file:   lex.llh
	new file:   lex.llm
	modified:   nil.llh
	modified:   nil.llm
	modified:   system.llh
	modified:   system_ll.llm

Renamed getchar_simple to getchar_checked and moved it to system since it is
essentially a better interface to stdio.  Also, system facilities should not
need to include nil.llh, and thus even the routines I write should all use
the C calling convention and C types.

Moved NextChar to a new lexer module.  Since it knows about which characters
are valid inputs, it belongs to the lexer and not the I/O routines in system.

Added a .gitignore file.

nil.llh now only includes c_defs.llh, not system.llh.