v0.0.2--Working lexer, friendlified input loop.
On branch master

	modified:   README
	modified:   lex.llh
	modified:   lex.llm
	modified:   nil.llm
	modified:   system.llh
	modified:   system_ll.llm

Version bumped up to 0.0.2, as this is a feature release.  The major
feature is actual lexing of the input line with a proper state machine
in lex::NextLexeme().  It will require additions for things like a
number type, but the basic code should not need replacing.  (But now
that I've said that publicly....)

Supporting changes and cleanup:

system:

    * Added a nicer wrapper for ungetc since the lexer needs it.

lex:

    * ClearLexemeBuffer, AppendLexemeChar, and GetLexemePtr implement a
      buffer for accumulating a lexeme character-by-character.

    * Eliminated error reporting in the lexer.

nil.llm

    * All error reporting is now at top level where it should be.

    * Added a welcome message to the input loop.

    * Friendlified the input behavior a bit.