Now with real lists, and the string list uses it.
Changes to be committed:

	modified:   README
	modified:   exp.llh
	modified:   exp.llm
	modified:   nil.llm
	modified:   system.llh

I hope this is the last release which mostly implements internal
features not obvious to the user.  The major advance is the
implementation of real lisp lists.

The list of unique strings now uses these lisp-native lists, which
required adding enough functionality that we now have implementations of
some of the minimal set of lisp primitives in Paul Graham's "The Roots
of Lisp" paper (http://www.paulgraham.com/rootsoflisp.html).  cons, car,
and cdr are fully implemented and atom only requires a wrapper to lift
the i1 boolean that LLVM uses to the t/nil values that lisp uses for
booleans.

This required adding a list type to PrintExp() (which turns a simple
function into a family of mutually recursive functions), so the command
interpreter uses that to show the list of known strings.

More nitty-gritty details that came along with that work are (of course)
a type for cons cells and a preprocessor symbol for the special value of
nil.

The tag field implementation was improved by using a three-bit int type
and the LLVM conversion intrinsics instead of the bit hackery I'd use in
C.  It probably doesn't change the performance (and may not even change
the generated machine code) but does allow LLVM to do more type
checking.