Broke out the memory module & fixed NIL.
Changes to be committed:

	modified:   Makefile
	modified:   exp.llh
	modified:   exp.llm
	new file:   memory.llh
	new file:   memory.llm
	modified:   nil.llh
	modified:   nil.llm

Moved the string uniqueifying code to a new memory module.

Changed the representation of NIL from null tagged with CELL_TAG to one
tagged with SYMBOL_TAG, as it makes the most common test (is this a cell
or an atom?) in the parser a single test.