Asserts now always fire--use NDEBUG manually.
Changes to be committed:

	modified:   exp.llm
	modified:   lex.llm
	modified:   memory.llm
	modified:   system.llh
	modified:   system_ll.llm

The old assert system was faulty because it tried to imitate the C
facility.  But C has nested expressions and LLVM doesn't, so in practice
one-line asserts are very rare.  Generally one has to do a compare
first, which means you have to put in the #ifndef NDEBUG guard anyway.
Also, a few times I had to call assert_impl by hand because I didn't
want it taken out.  So now assert's always fire no matter the value of
NDEBUG, and you just do it by hand.

Also, there is an assert_false() macro which at one point I needed and
might need again to save having to do a pointless xor to negate the
value being tested.