Release cleanup.
Changes to be committed:

	modified:   README
	modified:   nil.llm

From the README:

    v0.0.4.p1 * Cleanup release prepatory to parsing--the only visible
                change is the (+) notation for a new string added to
                the list.

[...]

Note that this is the same user-visible behavior as in v0.0.4 except for
the (+) notation (instead of (new)).  The major changes from 0.0.4 are
all to make implementing the parser easier--the lexer returns structs
and the error codes are collapsed into a single exception token.  In
addition, the code has received a project-wide cleanup.