More test functions, name cleanup.
Main changes: most of the test functions now are prefixed with 'Is,' and
there are a lot more of them.  All the bare comparisons with NIL_VALUE
in the rest of the code should be gone--we will depend on LLVM to inline
them when we care to optimize.

Work on the environment for eval, but in progress and not yet
functional.