Environment now has nested contexts.
The main addition here is that the environment is now a list (actually
stack) of nested contexts, each of which is an association list.

Added the IsList() test, even though it wasn't needed.

Added ExceptionMsg() to the expression module because it will be needed
for more sophisticated error reporting someday.